projects/quickshop-mvp-test4/backend/requirements.txt
# Core Web Framework
fastapi==0.95.1
uvicorn==0.22.0

# API Integrations
stripe==5.4.0
sendgrid==6.9.7
boto3==1.26.137

# Analytics
google-analytics-data==0.14.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.10
psycopg2-binary==2.9.6
alembic==1.10.3

# Validation
pydantic==1.10.7
pydantic-settings==2.0.3

# HTTP Requests
requests==2.30.0

# Environment Management
python-dotenv==1.0.0

# Async Support
asyncio==3.4.3

# Monitoring and Logging
sentry-sdk==1.22.2

# Testing
pytest==7.3.1
httpx==0.24.0

# Performance
redis==4.5.4

# Type Checking
mypy==1.3.0
types-requests==2.30.0